requirements-public.txt
# Python 3.12 호환 공개 패키지 요구사항
# 모든 의존성은 Python 3.12+ 호환성을 보장합니다

# === 웹 프레임워크 및 API ===
fastapi>=0.104.1
uvicorn>=0.24.0
starlette>=0.27.0
aiohttp>=3.9.1

# === 데이터베이스 ===
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
redis>=5.0.1
aioredis>=2.0.1
motor>=3.3.2  # MongoDB 비동기 드라이버
prisma>=0.11.0
elasticsearch>=8.11.1

# === 인증 및 보안 ===
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
pyotp>=2.9.0
firebase_admin>=6.3.0

# === Git 및 유틸리티 ===
GitPython>=3.1.41
setuptools>=69.0.2
wheel>=0.42.0
geopy>=2.4.1            # 지리 정보 처리
psutil>=5.9.6
requests>=2.31.0
importlib-metadata>=7.0.0
typing-extensions>=4.8.0
packaging>=23.2
urllib3>=2.1.0
httpx>=0.25.2  # HTTP 클라이언트
pydantic_settings>=2.1.0
routes>=2.5.1
python-dotenv>=1.0.0  # dotenv 패키지명 수정
structlog>=23.2.0

# === 데이터 처리 및 시각화 ===
numpy>=1.26.2
pandas>=2.1.3
matplotlib>=3.8.2
plotly>=5.18.0
tabulate>=0.9.0
msgpack>=1.0.7
bson>=0.5.10
orjson>=3.9.10
markdownify>=0.11.6
readabilipy>=0.2.0
protego>=0.3.0
python-json-logger>=2.0.7

# === 이메일 및 통신 ===
aiosmtplib>=2.0.2
fastapi_mail>=1.4.1
pywebpush>=1.14.0

# === 모니터링 및 로깅 ===
loguru>=0.7.2
prometheus_client>=0.17.1

# === 테스트 도구 ===
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.23.2
pytest-xdist>=3.5.0
freezegun>=1.2.0
pytest_html>=4.1.1
pip-audit>=2.6.1

# === 하위 버전 호환성 ===
pickle5==0.0.12; python_version < '3.8'  # Python 3.8 미만에서 호환성 제공